@@ -1,16 +1,44 @@
 <?php
 
+declare(strict_types=1);
+
+/**
+ *
+ */
+
 namespace App\Models;
 
 use App\Models\Locality\City;
+use App\Notifications\VerifyEmail;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Str;
 
-class User extends Authenticatable
+/**
+ * Class User
+ * @package App\Models
+ *
+ * @property string verify_token
+ */
+class User extends Authenticatable implements MustVerifyEmail
 {
-    use HasFactory, Notifiable;
+    use HasFactory;
+    use Notifiable;
+
+    /**
+     *
+     */
+    public const STATUS_WAIT = 'wait';
+
+    /**
+     *
+     */
+    public const STATUS_ACTIVE = 'active';
 
     /**
      * The attributes that are mass assignable.
@@ -25,6 +53,8 @@ class User extends Authenticatable
         'telephone',
         'photo',
         'city_id',
+        'verify_status',
+        'verify_token',
     ];
 
     /**
@@ -46,13 +76,82 @@ class User extends Authenticatable
         'email_verified_at' => 'datetime',
     ];
 
-    public function adverts()
+    /**
+     * @return BelongsToMany
+     */
+    public function adverts(): BelongsToMany
     {
         return $this->belongsToMany(Advert::class);
     }
 
-    public function city()
+    /**
+     * @return BelongsTo
+     */
+    public function city(): BelongsTo
     {
         return $this->belongsTo(City::class);
+    }
+
+    /**
+     *
+     */
+    public function sendEmailVerificationNotification(): void
+    {
+        $this->notify(new VerifyEmail());
+    }
+
+    /**
+     * Determine if the user has verified their email address.
+     *
+     * @return bool
+     */
+    public function hasVerifiedEmail(): bool
+    {
+        return $this->verify_status === self::STATUS_ACTIVE;
+    }
+
+    /**
+     *
+     */
+    public function verify(): void
+    {
+        $this->update([
+            'verify_token' => null,
+            'verify_status' => self::STATUS_ACTIVE,
+        ]);
+    }
+
+    /**
+     *
+     */
+    public function refreshToken(): void
+    {
+        $this->update([
+            'verify_token' => self::generateToken(),
+            'verify_status' => self::STATUS_WAIT,
+        ]);
+    }
+
+    /**
+     * @param array $data
+     * @return User
+     */
+    public static function make(array $data): User
+    {
+        return self::create([
+            'name' => $data['name'],
+            'email' => $data['email'],
+            'password' => Hash::make($data['password']),
+            'verify_status' => self::STATUS_WAIT,
+            'verify_token' => self::generateToken(),
+        ]);
+    }
+
+    /***
+     * @return string
+     */
+    protected static function generateToken(): string
+    {
+        return Str::random(20);
     }
 }
